mission2/code/tic-tac-toe-planner/pyproject.toml
[project]
name = "tic-tac-toe-planner"
version = "0.1.0"
description = "CLI that reads a tic-tac-toe board image, infers the state, and proposes the next move using Gemini Flash Lite 2.5."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.3",
    "obsws-python>=1.6.0",
    "opencv-python-headless>=4.10.0.84",
]

[project.scripts]
ttt-planner = "tic_tac_toe_planner:main"

[tool.uv]
package = true
